Create a specification table for a selected variable.

Returns estimates for a selected variable across the multiverse along with the universe specification information in a table. The resulting table can be used for spe_curve().

spec_summary(.mverse, var, conf.int = TRUE, conf.level = 0.95)

Arguments

.mverse

A mverse object.

var

A character specifying the variable of interest.

conf.int

Whether the table should include confidence intervals.

conf.level

The confidence level for the confidence level and is_significant.

Value

A spec_summary object that includes estimates and specification across the multiverse for the selected term(s). A boolean column is_significant indicates whether p.value for the universe is less than the specified significance level (1 - conf.level).

See also

Examples

femininity <- mutate_branch(
  1 * (MasFem > 6), 1 * (MasFem > mean(MasFem))
)
intensity <- mutate_branch(
  Minpressure_Updated_2014,
  Category,
  NDAM,
  HighestWindSpeed
)
model <- formula_branch(
  log(alldeaths + 1) ~ femininity,
  log(alldeaths + 1) ~ femininity * intensity
)
mv <- mverse(hurricane) %>%
  add_mutate_branch(femininity) %>%
  add_mutate_branch(intensity) %>%
  add_formula_branch(model) %>%
  lm_mverse()
spec_summary(mv, "femininity")
#> Specification table for variable: femininity at confidence intervals 0.95 
#> # A tibble: 16 × 13
#>    universe term       estimate p.value conf.low conf.high femininity_branch
#>    <fct>    <chr>         <dbl>   <dbl>    <dbl>     <dbl> <fct>            
#>  1 1        femininity   0.286    0.401   -0.386     0.957 femininity_1     
#>  2 2        femininity  15.5      0.247  -11.0      42.0   femininity_1     
#>  3 3        femininity   0.286    0.401   -0.386     0.957 femininity_1     
#>  4 4        femininity  -0.464    0.474   -1.75      0.820 femininity_1     
#>  5 5        femininity   0.286    0.401   -0.386     0.957 femininity_1     
#>  6 6        femininity   0.0552   0.849   -0.519     0.630 femininity_1     
#>  7 7        femininity   0.286    0.401   -0.386     0.957 femininity_1     
#>  8 8        femininity  -0.791    0.575   -3.58      2.00  femininity_1     
#>  9 9        femininity   0.331    0.326   -0.334     0.996 femininity_2     
#> 10 10       femininity  16.3      0.224  -10.1      42.6   femininity_2     
#> 11 11       femininity   0.331    0.326   -0.334     0.996 femininity_2     
#> 12 12       femininity  -0.510    0.429   -1.79      0.767 femininity_2     
#> 13 13       femininity   0.331    0.326   -0.334     0.996 femininity_2     
#> 14 14       femininity   0.0894   0.756   -0.479     0.658 femininity_2     
#> 15 15       femininity   0.331    0.326   -0.334     0.996 femininity_2     
#> 16 16       femininity  -0.938    0.504   -3.72      1.84  femininity_2     
#> # ℹ 6 more variables: intensity_branch <fct>, model_branch <fct>,
#> #   femininity_branch_code <fct>, intensity_branch_code <fct>,
#> #   model_branch_code <fct>, is_significant <lgl>