Display the multiverse table with results.

This method returns the multiverse table displaying all universes defined by the multiverse. Each row corresponds to a universe and each column represents a branch.

summary.lm_mverse returns the lm regression results across the multiverse.

summary.glm_mverse returns the glm regression results across the multiverse.

summary.glm.nb_mverse returns the MASS::glm.nb regression results across the multiverse.

# S3 method for mverse
summary(object, ...)

# S3 method for lm_mverse
summary(object, conf.int = TRUE, conf.level = 0.95, output = "estimates", ...)

# S3 method for glm_mverse
summary(object, conf.int = TRUE, conf.level = 0.95, output = "estimates", ...)

# S3 method for glm.nb_mverse
summary(object, conf.int = TRUE, conf.level = 0.95, output = "estimates", ...)

Arguments

object

a glm.nb_mverse object.

...

Ignored.

conf.int

When TRUE (default), the estimate output includes the confidence intervals.

conf.level

The confidence level of the confidence interval returned using conf.int = TRUE. Default value is 0.95.

output

The output of interest. The possible values are "estimates" ("e"), "df", "deviance" ("de"), and "aic" ("bic"). Alternatively, the first letters may be used. Default value is "estimates".

Value

a multiverse table as a tibble. a multiverse table as a tibble a multiverse table as a tibble a multiverse table as a tibble

See also

Examples

# Define a mutate branch.
hurricane_strength <- mutate_branch(
  # damage vs. wind speed vs.pressure
  NDAM,
  HighestWindSpeed,
  Minpressure_Updated_2014,
  # Standardized versions
  scale(NDAM),
  scale(HighestWindSpeed),
  -scale(Minpressure_Updated_2014),
)
# Create a mverse and add the branch.
mv <- create_multiverse(hurricane) %>%
  add_mutate_branch(hurricane_strength)
# Display the multiverse table while
# defining a multiverse provides a way
# to inspect the universes created.
summary(mv)
#> # A tibble: 6 × 2
#>   universe hurricane_strength_branch       
#>   <fct>    <fct>                           
#> 1 1        NDAM                            
#> 2 2        HighestWindSpeed                
#> 3 3        Minpressure_Updated_2014        
#> 4 4        scale(NDAM)                     
#> 5 5        scale(HighestWindSpeed)         
#> 6 6        -scale(Minpressure_Updated_2014)
# Define a filter branch.
hurricane_outliers <- filter_branch(
  !Name %in% c("Katrina", "Audrey", "Andrew"),
  !Name %in% c("Katrina"),
  !Name %in% c("Katrina"),
  TRUE # include all
)
# Add the additional branch and inspect.
mv <- add_filter_branch(mv, hurricane_outliers)
summary(mv)
#> # A tibble: 24 × 3
#>    universe hurricane_strength_branch hurricane_outliers_branch                 
#>    <fct>    <fct>                     <fct>                                     
#>  1 1        NDAM                      "!Name %in% c(\"Katrina\", \"Audrey\", \"…
#>  2 2        NDAM                      "!Name %in% c(\"Katrina\")"               
#>  3 3        NDAM                      "!Name %in% c(\"Katrina\")"               
#>  4 4        NDAM                      "TRUE"                                    
#>  5 5        HighestWindSpeed          "!Name %in% c(\"Katrina\", \"Audrey\", \"…
#>  6 6        HighestWindSpeed          "!Name %in% c(\"Katrina\")"               
#>  7 7        HighestWindSpeed          "!Name %in% c(\"Katrina\")"               
#>  8 8        HighestWindSpeed          "TRUE"                                    
#>  9 9        Minpressure_Updated_2014  "!Name %in% c(\"Katrina\", \"Audrey\", \"…
#> 10 10       Minpressure_Updated_2014  "!Name %in% c(\"Katrina\")"               
#> # … with 14 more rows
# Define mutate branches.
hurricane_strength <- mutate_branch(
  # damage vs. wind speed vs.pressure
  NDAM,
  HighestWindSpeed,
  Minpressure_Updated_2014,
  # Standardized versions
  scale(NDAM),
  scale(HighestWindSpeed),
  -scale(Minpressure_Updated_2014),
)
y <- mutate_branch(
  alldeaths, log(alldeaths + 1)
)
# Define a filter branch.
hurricane_outliers <- filter_branch(
  !Name %in% c("Katrina", "Audrey", "Andrew"),
  !Name %in% c("Katrina"),
  !Name %in% c("Katrina"),
  TRUE # include all
)
# Define a formula branch.
model_specifications <- formula_branch(
  y ~ femininity,
  y ~ femininity + hurricane_strength,
  y ~ femininity * hurricane_strength
)
# Create a mverse, add the branches, and fit lm models.
mv <- create_multiverse(hurricane) %>%
  add_filter_branch(hurricane_outliers) %>%
  add_mutate_branch(hurricane_strength, y) %>%
  add_formula_branch(model_specifications) %>%
  lm_mverse()
#> Warning: error in default universe
#> Error in eval(predvars, data, env) : object 'femininity' not found
#> pkgdown::deploy_to_branch -> build_site_github_pages -> build_site -> build_site_local -> build_reference -> purrr::map -> .f -> withCallingHandlers -> data_reference_topic -> run_examples -> highlight_examples -> downlit::evaluate_and_highlight -> evaluate::evaluate -> evaluate_call -> timing_fn -> handle -> try - f, silent = TRUE) -> withCallingHandlers -> withVisible -> eval_with_user_handlers -> eval -> eval -> create_multiverse -> lm_mverse -> multiverse::inside -> execute_universe -> mapply ->  -> execute_code_from_universe -> tryStack -> lapply -> FUN -> FUN -> stats::lm -> eval -> eval -> stats::model.frame -> model.frame.default -> eval -> eval -> eval(predvars, data, env)
# Display the multiverse table with estimated coefficients.
summary(mv)
#> Warning: error in default universe
#> Error in summary(model) : object 'model' not found
#> pkgdown::deploy_to_branch -> build_site_github_pages -> build_site -> build_site_local -> build_reference -> purrr::map -> .f -> withCallingHandlers -> data_reference_topic -> run_examples -> highlight_examples -> downlit::evaluate_and_highlight -> evaluate::evaluate -> evaluate_call -> timing_fn -> handle -> try - f, silent = TRUE) -> withCallingHandlers -> withVisible -> eval_with_user_handlers -> eval -> eval -> summary -> summary.lm_mverse -> multiverse::inside -> execute_universe -> mapply ->  -> execute_code_from_universe -> tryStack -> lapply -> FUN -> FUN -> summary -> summary(model)
#> # A tibble: 144 × 6
#>    universe hurricane_outliers… hurricane_stren… y_branch model_specifica… out  
#>    <fct>    <fct>               <fct>            <fct>    <fct>            <lgl>
#>  1 1        "!Name %in% c(\"Ka… NDAM             alldeat… y ~ femininity   NA   
#>  2 2        "!Name %in% c(\"Ka… NDAM             alldeat… y ~ femininity … NA   
#>  3 3        "!Name %in% c(\"Ka… NDAM             alldeat… y ~ femininity … NA   
#>  4 4        "!Name %in% c(\"Ka… NDAM             log(all… y ~ femininity   NA   
#>  5 5        "!Name %in% c(\"Ka… NDAM             log(all… y ~ femininity … NA   
#>  6 6        "!Name %in% c(\"Ka… NDAM             log(all… y ~ femininity … NA   
#>  7 7        "!Name %in% c(\"Ka… HighestWindSpeed alldeat… y ~ femininity   NA   
#>  8 8        "!Name %in% c(\"Ka… HighestWindSpeed alldeat… y ~ femininity … NA   
#>  9 9        "!Name %in% c(\"Ka… HighestWindSpeed alldeat… y ~ femininity … NA   
#> 10 10       "!Name %in% c(\"Ka… HighestWindSpeed log(all… y ~ femininity   NA   
#> # … with 134 more rows
# Define mutate branches.
hurricane_strength <- mutate_branch(
  # damage vs. wind speed vs.pressure
  NDAM,
  HighestWindSpeed,
  Minpressure_Updated_2014,
  # Standardized versions
  scale(NDAM),
  scale(HighestWindSpeed),
  -scale(Minpressure_Updated_2014),
)
y <- mutate_branch(
  alldeaths, log(alldeaths + 1)
)
# Define a filter branch.
hurricane_outliers <- filter_branch(
  !Name %in% c("Katrina", "Audrey", "Andrew"),
  !Name %in% c("Katrina"),
  !Name %in% c("Katrina"),
  TRUE # include all
)
# Define a formula branch.
model_specifications <- formula_branch(
  y ~ femininity,
  y ~ femininity + hurricane_strength,
  y ~ femininity * hurricane_strength
)
# Define a family branch.
model_distributions <- family_branch(gaussian, poisson)
# Create a mverse, add the branches, and fit glm models.
mv <- create_multiverse(hurricane) %>%
  add_filter_branch(hurricane_outliers) %>%
  add_mutate_branch(hurricane_strength, y) %>%
  add_formula_branch(model_specifications) %>%
  add_family_branch(model_distributions) %>%
  glm_mverse()
#> Warning: error in default universe
#> Error in eval(predvars, data, env) : object 'femininity' not found
#> pkgdown::deploy_to_branch -> build_site_github_pages -> build_site -> build_site_local -> build_reference -> purrr::map -> .f -> withCallingHandlers -> data_reference_topic -> run_examples -> highlight_examples -> downlit::evaluate_and_highlight -> evaluate::evaluate -> evaluate_call -> timing_fn -> handle -> try - f, silent = TRUE) -> withCallingHandlers -> withVisible -> eval_with_user_handlers -> eval -> eval -> create_multiverse -> glm_mverse -> multiverse::inside -> execute_universe -> mapply ->  -> execute_code_from_universe -> tryStack -> lapply -> FUN -> FUN -> stats::glm -> eval -> eval -> stats::model.frame -> model.frame.default -> eval -> eval -> eval(predvars, data, env)
# Display the multiverse table with estimated coefficients.
summary(mv)
#> Warning: error in default universe
#> Error in summary(model) : object 'model' not found
#> pkgdown::deploy_to_branch -> build_site_github_pages -> build_site -> build_site_local -> build_reference -> purrr::map -> .f -> withCallingHandlers -> data_reference_topic -> run_examples -> highlight_examples -> downlit::evaluate_and_highlight -> evaluate::evaluate -> evaluate_call -> timing_fn -> handle -> try - f, silent = TRUE) -> withCallingHandlers -> withVisible -> eval_with_user_handlers -> eval -> eval -> summary -> summary.glm_mverse -> multiverse::inside -> execute_universe -> mapply ->  -> execute_code_from_universe -> tryStack -> lapply -> FUN -> FUN -> summary -> summary(model)
#> # A tibble: 288 × 7
#>    universe hurricane_outliers_branch hurricane_stren… y_branch model_specifica…
#>    <fct>    <fct>                     <fct>            <fct>    <fct>           
#>  1 1        "!Name %in% c(\"Katrina\… NDAM             alldeat… y ~ femininity  
#>  2 2        "!Name %in% c(\"Katrina\… NDAM             alldeat… y ~ femininity  
#>  3 3        "!Name %in% c(\"Katrina\… NDAM             alldeat… y ~ femininity …
#>  4 4        "!Name %in% c(\"Katrina\… NDAM             alldeat… y ~ femininity …
#>  5 5        "!Name %in% c(\"Katrina\… NDAM             alldeat… y ~ femininity …
#>  6 6        "!Name %in% c(\"Katrina\… NDAM             alldeat… y ~ femininity …
#>  7 7        "!Name %in% c(\"Katrina\… NDAM             log(all… y ~ femininity  
#>  8 8        "!Name %in% c(\"Katrina\… NDAM             log(all… y ~ femininity  
#>  9 9        "!Name %in% c(\"Katrina\… NDAM             log(all… y ~ femininity …
#> 10 10       "!Name %in% c(\"Katrina\… NDAM             log(all… y ~ femininity …
#> # … with 278 more rows, and 2 more variables: model_distributions_branch <fct>,
#> #   out <lgl>
# Define mutate branches.
hurricane_strength <- mutate_branch(
  # damage vs. wind speed vs.pressure
  NDAM,
  HighestWindSpeed,
  Minpressure_Updated_2014,
  # Standardized versions
  scale(NDAM),
  scale(HighestWindSpeed),
  -scale(Minpressure_Updated_2014),
)
y <- mutate_branch(
  alldeaths, log(alldeaths + 1)
)
# Define a filter branch.
hurricane_outliers <- filter_branch(
  !Name %in% c("Katrina", "Audrey", "Andrew"),
  !Name %in% c("Katrina"),
  !Name %in% c("Katrina"),
  TRUE # include all
)
# Define a formula branch.
model_specifications <- formula_branch(
  y ~ femininity,
  y ~ femininity + hurricane_strength,
  y ~ femininity * hurricane_strength
)
# Create a mverse, add the branches, and fit glm models.
mv <- create_multiverse(hurricane) %>%
  add_filter_branch(hurricane_outliers) %>%
  add_mutate_branch(hurricane_strength, y) %>%
  add_formula_branch(model_specifications) %>%
  glm.nb_mverse()
#> Warning: error in default universe
#> Error in eval(predvars, data, env) : object 'femininity' not found
#> pkgdown::deploy_to_branch -> build_site_github_pages -> build_site -> build_site_local -> build_reference -> purrr::map -> .f -> withCallingHandlers -> data_reference_topic -> run_examples -> highlight_examples -> downlit::evaluate_and_highlight -> evaluate::evaluate -> evaluate_call -> timing_fn -> handle -> try - f, silent = TRUE) -> withCallingHandlers -> withVisible -> eval_with_user_handlers -> eval -> eval -> create_multiverse -> glm.nb_mverse -> multiverse::inside -> execute_universe -> mapply ->  -> execute_code_from_universe -> tryStack -> lapply -> FUN -> FUN -> MASS::glm.nb -> eval.parent -> eval -> eval -> stats::model.frame -> model.frame.default -> eval -> eval -> eval(predvars, data, env)
# Display the multiverse table with estimated coefficients.
summary(mv)
#> Warning: error in default universe
#> Error in summary(model) : object 'model' not found
#> pkgdown::deploy_to_branch -> build_site_github_pages -> build_site -> build_site_local -> build_reference -> purrr::map -> .f -> withCallingHandlers -> data_reference_topic -> run_examples -> highlight_examples -> downlit::evaluate_and_highlight -> evaluate::evaluate -> evaluate_call -> timing_fn -> handle -> try - f, silent = TRUE) -> withCallingHandlers -> withVisible -> eval_with_user_handlers -> eval -> eval -> summary -> summary.glm.nb_mverse -> multiverse::inside -> execute_universe -> mapply ->  -> execute_code_from_universe -> tryStack -> lapply -> FUN -> FUN -> summary -> summary(model)
#> # A tibble: 144 × 6
#>    universe hurricane_outliers… hurricane_stren… y_branch model_specifica… out  
#>    <fct>    <fct>               <fct>            <fct>    <fct>            <lgl>
#>  1 1        "!Name %in% c(\"Ka… NDAM             alldeat… y ~ femininity   NA   
#>  2 2        "!Name %in% c(\"Ka… NDAM             alldeat… y ~ femininity … NA   
#>  3 3        "!Name %in% c(\"Ka… NDAM             alldeat… y ~ femininity … NA   
#>  4 4        "!Name %in% c(\"Ka… NDAM             log(all… y ~ femininity   NA   
#>  5 5        "!Name %in% c(\"Ka… NDAM             log(all… y ~ femininity … NA   
#>  6 6        "!Name %in% c(\"Ka… NDAM             log(all… y ~ femininity … NA   
#>  7 7        "!Name %in% c(\"Ka… HighestWindSpeed alldeat… y ~ femininity   NA   
#>  8 8        "!Name %in% c(\"Ka… HighestWindSpeed alldeat… y ~ femininity … NA   
#>  9 9        "!Name %in% c(\"Ka… HighestWindSpeed alldeat… y ~ femininity … NA   
#> 10 10       "!Name %in% c(\"Ka… HighestWindSpeed log(all… y ~ femininity   NA   
#> # … with 134 more rows