Number of cards given for each referee-player pair in soccer.

A dataset containing card counts between 2,053 soccer players playing in the first male divisions of England, Germany, France, and Spain in the 2012-2013 season and 3,147 referees that these players played under in professional matches. The dataset contains other covariates including 2 independent skin tone ratings per player. Each line represents a player-referee pair.

soccer

Format

A data frame with 146,028 rows and 26 variables:

playerShort

short player ID

player

player name

club

player club

leagueCountry

country of player club ( England, Germany, France, and Spain)

birthday

player birthday

height

player height (in cm)

weight

player weight (in kg)

position

detailed player position

games

number of games in the player-referee dyad

victories

victories in the player-referee dyad

ties

ties in the player-referee dyad

defeats

losses in the player-referee dyad

goals

goals scored by a player in the player-referee dyad

yellowCards

number of yellow cards player received from referee

yellowReds

number of yellow-red cards player received from referee

redCards

number of red cards player received from referee

rater1

skin rating of photo by rater 1 (5-point scale ranging from “very light skin” to “very dark skin”)

rater2

skin rating of photo by rater 2 (5-point scale ranging from “very light skin” to “very dark skin”)

refNum

unique referee ID number (referee name removed for anonymizing purposes)

refCountry

unique referee country ID number (country name removed for anonymizing purposes)

meanIAT

mean implicit bias score (using the race IAT) for referee country, higher values correspond to faster white | good, black | bad associations

nIAT

sample size for race IAT in that particular country

seIAT

standard error for mean estimate of race IAT

meanExp

mean explicit bias score (using a racial thermometer task) for referee country, higher values correspond to greater feelings of warmth toward whites versus blacks

nExp

sample size for explicit bias in that particular country

seExp

standard error for mean estimate of explicit bias measure

Source

Silberzahn, R., Uhlmann, E. L., Martin, D. P., Anselmi, P., Aust, F., Awtrey, E. C., … Nosek, B. A. (2018, August 24). Many analysts, one dataset: Making transparent how variations in analytical choices affect results. Retrieved from https://osf.io/gvm2z/

Details

The skin colour of each player was rated by two independent raters, rater1 and rater2, and the 5-point scale values were scaled to 0 to 1 - i.e., 0, 0.25, 0.5, 0.75, 1.